Los enemigos dañan al jugador
Modificaciones a los scrip hechos por facundo para que los enemigos hagan daño al jugador.
- Del enemigo mele: Agregado el evento ontriggerevent. Oncollision sin modificaciones.
- Del enemigo rango: Modificado el tipo de proyectil que dispara, se tuvo que modificar la parte del código donde se instancia la bala para que cree una acorde al tipo pedido. La bala original, NO FUE eliminada, ni modificada, se creo una nueva.

+++ Rise to Stardom/Assets/Scripts/Player/DispararJugador.cs
@@ -19,7 +19,7 @@ public class DispararJugador : MonoBehaviour
         firepointHandler = new FirepointHandler(firePoint, transform, firePointDistance);
     }
 
-    public void Disparar(RaycastHit mouse, int damage, int speed)
+    public void Disparar(RaycastHit mouse, float damage, float speed)
     {
 
         Vector3 pointerPosition = mouse.point;
